feat: native C615 V4L2 controls with hardware zoom, remove digital zoom crop/scale, simplify vflip filter
- Rewrite CameraService.cs: native 0-255 V4L2 values, hardware zoom_absolute (1-5) via v4l2-ctl, exposure auto/manual level mapping, white balance auto/temperature presets
- PtzController.cs: sharpness/contrast/saturation parameter types double -> int
- dock.html: slider ranges and defaults for C615 (zoom 1-5, focus 0-255, exposure 0-8, wb 0-6, native 0-255 for brightness/contrast/saturation/sharpness), wbModes 7 C615 names, exposure display Auto/Level N
- BuildFfmpegCaptureArgs: removed digital zoom crop/scale, fixed vflip syntax to plain filter (no enable= expression)
- ApplyV4l2Controls: simplified mapping, no more double-to-int conversion
- VERSION bump to 1.1.3
- .gitignore: add publish-linux/, publish-fd/

+++ pi stream cam/Controllers/PtzController.cs
@@ -122,7 +122,7 @@ public class PtzController : ControllerBase
     }
 
     [HttpPost("sharpness/{value}")]
-    public IActionResult SetSharpness(double value)
+    public IActionResult SetSharpness(int value)
     {
         _cameraService.SetSharpness(value);
         return Ok(new { sharpness = _cameraService.Sharpness });
@@ -136,14 +136,14 @@ public class PtzController : ControllerBase
     }
 
     [HttpPost("contrast/{value}")]
-    public IActionResult SetContrast(double value)
+    public IActionResult SetContrast(int value)
     {
         _cameraService.SetContrast(value);
         return Ok(new { contrast = _cameraService.Contrast });
     }
 
     [HttpPost("saturation/{value}")]
-    public IActionResult SetSaturation(double value)
+    public IActionResult SetSaturation(int value)
     {
         _cameraService.SetSaturation(value);
         return Ok(new { saturation = _cameraService.Saturation });
